Move common mocks into a @deephaven/mocks package
