+++ packages/code-studio/src/__mocks__/lodash.debounce.js
@@ -1,0 +1,1 @@
+export { default } from '@deephaven/mocks/lodash.debounce';
