+++ packages/code-studio/src/__mocks__/monaco-editor-empty.js
@@ -1,1 +1,2 @@
 // An empty mock for all the feature imports we need from monaco
+// Used from the `moduleNameMapper` section in package.json
