+++ packages/code-studio/src/__mocks__/pouchdb-browser.js
@@ -1,12 +1,1 @@
-const PouchDB = jest.fn().mockImplementation(() => ({
-  changes: jest.fn(() => ({
-    on: jest.fn(() => {}),
-  })),
-  createIndex: jest.fn(),
-  find: jest.fn(() => ({
-    docs: [],
-  })),
-}));
-PouchDB.plugin = jest.fn();
-
-export default PouchDB;
+export { default } from '@deephaven/mocks/pouchdb-browser';
