+++ packages/code-studio/src/__mocks__/react-virtualized-auto-sizer.js
@@ -1,1 +1,1 @@
-export default ({ children }) => children({ height: 500, width: 500 });
+export { default } from '@deephaven/mocks/react-virtualized-auto-sizer';
