+++ packages/iris-grid/src/__mocks__/lodash.throttle.js
@@ -1,0 +1,1 @@
+export { default } from '@deephaven/mocks/lodash.throttle';
