+++ packages/mocks/.gitignore
@@ -1,0 +1,2 @@
+# See https://help.github.com/ignore-files/ for more about ignoring files.
+/node_modules
