+++ packages/mocks/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@deephaven/mocks",
+  "version": "0.1.0",
+  "description": "Deephaven Mocks for common libraries",
+  "repository": "https://github.com/deephaven/web-client-ui",
+  "author": "Deephaven Data Labs LLC",
+  "license": "Apache-2.0",
+  "main": "./index.js",
+  "peerDependencies": { "jest": "26.x" },
+  "publishConfig": {
+    "access": "public"
+  }
+}
