+++ packages/mocks/react-virtualized-auto-sizer.js
@@ -1,0 +1,1 @@
+export default ({ children }) => children({ height: 500, width: 500 });
